# config/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E-ink home display server",
    "description": "What displays and monitors ask of the server that serves the dashboard image.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/plan": {
      "get": {
        "tags": [
          "displays"
        ],
        "summary": "Tells a display which render the image is, whether it is stale, and when to ask again.",
        "operationId": "plan",
        "parameters": [
          {
            "name": "device",
            "in": "query",
            "description": "The display's name, so the server can tell displays apart (battery, last seen, which image it was sent).\nA name that is not usable is ignored and the request is served all the same. Over HTTPS the display is\nnamed by its certificate and this is ignored.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "pattern": "^[A-Za-z0-9._-]+$"
            },
            "example": "reterminal-e1003-a1b2c3"
          },
          {
            "name": "have",
            "in": "query",
            "description": "The version of the image the display already shows (the `version` of an earlier plan), so the answer can\nsay the image is unchanged and the display can skip the download and the refresh.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1791463200
          },
          {
            "name": "battery_mv",
            "in": "query",
            "description": "Battery voltage, in millivolts. Ignored unless a whole number from 2000 to 5000.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 3712
          },
          {
            "name": "battery_pct",
            "in": "query",
            "description": "Battery charge, as a percentage. Ignored unless a whole number from 0 to 100.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 47
          },
          {
            "name": "battery_state",
            "in": "query",
            "description": "How the display rates its own battery. The display decides: the server only passes it on.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/BatteryState"
            }
          },
          {
            "name": "failed_wakes",
            "in": "query",
            "description": "How many wakes in a row have failed, for any reason.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 2
          },
          {
            "name": "rssi",
            "in": "query",
            "description": "Wi-Fi signal strength, in dBm. Ignored unless a whole number from -127 to -1.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": -71
          },
          {
            "name": "last_failure",
            "in": "query",
            "description": "Why the wake before this one failed, if it did.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FailureReason"
            }
          },
          {
            "name": "last_wake_s",
            "in": "query",
            "description": "How long the wake before this one was awake, in seconds: what costs battery. Ignored unless a whole number from 0 to 600.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 24
          },
          {
            "name": "last_tls_ms",
            "in": "query",
            "description": "How long the display's first TLS handshake of the wake before this one took, in milliseconds, as the\ndisplay timed it (the server also times them, from its side: `connection` in `/status`).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 1100
          },
          {
            "name": "last_heap_min",
            "in": "query",
            "description": "The least free heap the display had during the wake before this one, in bytes: how close the handshake\ncame to running it out.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "example": 61440
          },
          {
            "name": "fw",
            "in": "query",
            "description": "The version of the firmware the display runs, as released (letters, digits, `.`, `-` and `_`, at most 32).",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "0.2.0"
          }
        ],
        "responses": {
          "200": {
            "description": "The plan for the image there is.",
            "headers": {
              "cache-control": {
                "schema": {
                  "type": "string"
                },
                "description": "Always `no-store`."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "What a display is told after it asks: which image is current, whether it is out of date, and when to come back.",
                  "required": [
                    "version",
                    "changed",
                    "stale",
                    "pending",
                    "next_seconds",
                    "age_seconds",
                    "timezone",
                    "utc_offset_seconds"
                  ],
                  "properties": {
                    "version": {
                      "type": "integer",
                      "format": "int32",
                      "description": "When the image was rendered, in seconds since 1970. Newer images have larger versions.",
                      "examples": [
                        1791463200
                      ],
                      "minimum": 0
                    },
                    "changed": {
                      "type": "boolean",
                      "description": "False when the display said it already shows this version, so it can skip the download and refresh."
                    },
                    "stale": {
                      "type": "boolean",
                      "description": "A scheduled render is long overdue, so the image is older than it should be."
                    },
                    "pending": {
                      "type": "boolean",
                      "description": "A render is due or running, so a new image is about to appear."
                    },
                    "next_seconds": {
                      "type": "integer",
                      "format": "int64",
                      "description": "How long the display should sleep before asking again.",
                      "examples": [
                        600
                      ],
                      "minimum": 0
                    },
                    "age_seconds": {
                      "type": "integer",
                      "format": "int64",
                      "description": "Time since the image was rendered.",
                      "examples": [
                        100
                      ],
                      "minimum": 0
                    },
                    "timezone": {
                      "type": "string",
                      "description": "The zone the server works in (an IANA name such as `Europe/London`), so the display needn't be told it in\nits own configuration, where it could disagree.",
                      "examples": [
                        "Europe/London"
                      ]
                    },
                    "utc_offset_seconds": {
                      "type": "integer",
                      "format": "int32",
                      "description": "How far ahead of UTC that zone is right now, in seconds (3600 in the UK in summer). A display has no\ntimezone database, so this is what it needs to show local time; it is right until the next clock change.",
                      "examples": [
                        3600
                      ]
                    }
                  }
                },
                "example": {
                  "age_seconds": 100,
                  "changed": true,
                  "next_seconds": 600,
                  "pending": false,
                  "stale": false,
                  "timezone": "Europe/London",
                  "utc_offset_seconds": 3600,
                  "version": 1791463200
                }
              }
            }
          },
          "404": {
            "description": "Nothing has been rendered yet, so there is no image to plan around.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "No image has been rendered yet"
              }
            }
          },
          "500": {
            "description": "The server could not work the plan out."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatteryState": {
        "type": "string",
        "enum": [
          "ok",
          "low",
          "empty"
        ]
      },
...
      "FailureReason": {
        "type": "string",
        "description": "Why a display's last wake failed, as it reports it on the next wake that gets through. A wake can't\nreport its own failure (it asks `/plan` before it knows), and a Wi-Fi or server failure can only be\ntold once the display is talking to the server again.",
        "enum": [
          "wifi",
          "server",
          "download",
          "memory",
          "timeout",
          "clock",
          "certificate",
          "approval",
          "unrecognised"
        ]
      },
...
    }
  },
...
}